chore: configure prettier husky, lint-staged
