+++ .prettierrc.js
@@ -1,7 +1,7 @@
 module.exports = {
-  arrowParens: 'avoid',
-  bracketSameLine: true,
-  bracketSpacing: false,
-  singleQuote: true,
-  trailingComma: 'all',
+	semi: true,
+	trailingComma: "none",
+	singleQuote: false,
+	printWidth: 80,
+	useTabs: true
 };
