+++ App.tsx
@@ -1,12 +1,12 @@
-import React from 'react';
-import {NavigationContainer} from '@react-navigation/native';
-import MyStack from './navigation/Stacks';
+import React from "react";
+import { NavigationContainer } from "@react-navigation/native";
+import MyStack from "./src/navigation/Stacks";
 
 const App = () => {
-  return (
-    <NavigationContainer>
-      <MyStack />
-    </NavigationContainer>
-  );
+	return (
+		<NavigationContainer>
+			<MyStack />
+		</NavigationContainer>
+	);
 };
 export default App;
