+++ app.json
@@ -1,4 +1,4 @@
 {
-  "name": "treetrackerWebmapCoreDemoRn",
-  "displayName": "treetrackerWebmapCoreDemoRn"
+	"name": "treetrackerWebmapCoreDemoRn",
+	"displayName": "treetrackerWebmapCoreDemoRn"
 }
