+++ ios/treetrackerWebmapCoreDemoRn/Images.xcassets/Contents.json
@@ -1,6 +1,6 @@
 {
-  "info" : {
-    "version" : 1,
-    "author" : "xcode"
-  }
+	"info": {
+		"version": 1,
+		"author": "xcode"
+	}
 }
