+++ package.json
@@ -1,41 +1,62 @@
 {
-  "name": "treetrackerWebmapCoreDemoRn",
-  "version": "0.0.1",
-  "private": true,
-  "scripts": {
-    "android": "react-native run-android",
-    "ios": "react-native run-ios",
-    "lint": "eslint .",
-    "start": "react-native start",
-    "test": "jest"
-  },
-  "dependencies": {
-    "@react-navigation/native": "^6.1.6",
-    "@react-navigation/native-stack": "^6.9.12",
-    "react": "18.2.0",
-    "react-native": "0.71.7",
-    "react-native-safe-area-context": "^4.5.2",
-    "react-native-screens": "^3.20.0",
-    "react-native-webview": "^12.0.2"
-  },
-  "devDependencies": {
-    "@babel/core": "^7.20.0",
-    "@babel/preset-env": "^7.20.0",
-    "@babel/runtime": "^7.20.0",
-    "@react-native-community/eslint-config": "^3.2.0",
-    "@tsconfig/react-native": "^2.0.2",
-    "@types/jest": "^29.2.1",
-    "@types/react": "^18.0.24",
-    "@types/react-test-renderer": "^18.0.0",
-    "babel-jest": "^29.2.1",
-    "eslint": "^8.19.0",
-    "jest": "^29.2.1",
-    "metro-react-native-babel-preset": "0.73.9",
-    "prettier": "^2.4.1",
-    "react-test-renderer": "18.2.0",
-    "typescript": "4.8.4"
-  },
-  "jest": {
-    "preset": "react-native"
-  }
+	"name": "treetrackerWebmapCoreDemoRn",
+	"version": "0.0.1",
+	"private": true,
+	"scripts": {
+		"android": "react-native run-android",
+		"ios": "react-native run-ios",
+		"lint": "eslint .",
+		"prettier-format": "prettier --config .prettierrc.js 'src/**/*.{js,jsx,ts,tsx}' --write",
+		"start": "react-native start",
+		"test": "jest",
+		"prepare": "husky install"
+	},
+	"dependencies": {
+		"-": "^0.0.1",
+		"@react-navigation/native": "^6.1.6",
+		"@react-navigation/native-stack": "^6.9.12",
+		"@typescript-eslint/eslint-plugin": "^5.59.2",
+		"D": "^1.0.0",
+		"eslint-config-prettier": "^8.8.0",
+		"eslint-plugin-detox": "^1.0.0",
+		"react": "18.2.0",
+		"react-native": "0.71.7",
+		"react-native-safe-area-context": "^4.5.2",
+		"react-native-screens": "^3.20.0",
+		"react-native-vector-icons": "^9.2.0",
+		"react-native-webview": "^12.0.2"
+	},
+	"devDependencies": {
+		"@babel/core": "^7.20.0",
+		"@babel/eslint-parser": "^7.21.8",
+		"@babel/preset-env": "^7.20.0",
+		"@babel/preset-react": "^7.18.6",
+		"@babel/runtime": "^7.20.0",
+		"@react-native-community/eslint-config": "^3.2.0",
+		"@tsconfig/react-native": "^2.0.2",
+		"@types/jest": "^29.2.1",
+		"@types/react": "^18.0.24",
+		"@types/react-native-vector-icons": "^6.4.13",
+		"@types/react-test-renderer": "^18.0.0",
+		"babel-jest": "^29.2.1",
+		"eslint": "^8.39.0",
+		"eslint-config-airbnb": "^19.0.4",
+		"eslint-plugin-jest": "^27.2.1",
+		"husky": "^8.0.3",
+		"jest": "^29.2.1",
+		"lint-staged": "^13.2.2",
+		"metro-react-native-babel-preset": "0.73.9",
+		"prettier": "^2.8.8",
+		"react-test-renderer": "18.2.0",
+		"typescript": "4.8.4"
+	},
+	"jest": {
+		"preset": "react-native"
+	},
+	"lint-staged": {
+		"*.{js,jsx,ts,tsx}": [
+			"prettier --write",
+			"eslint --fix"
+		]
+	}
 }
