+++ react-native-config.js
@@ -1,0 +1,9 @@
+module.exports = {
+	dependencies: {
+		"react-native-vector-icons": {
+			platforms: {
+				ios: null
+			}
+		}
+	}
+};
