+++ src/navigation/Stacks.tsx
@@ -1,0 +1,19 @@
+import React from "react";
+import { createNativeStackNavigator } from "@react-navigation/native-stack";
+import HomeScreen from "../screens/HomeScreen";
+
+const Stack = createNativeStackNavigator();
+
+function MyStack() {
+	return (
+		<Stack.Navigator
+			screenOptions={{
+				headerShown: false
+			}}
+		>
+			<Stack.Screen name="Home" component={HomeScreen} />
+		</Stack.Navigator>
+	);
+}
+
+export default MyStack;
