+++ tsconfig.json
@@ -1,3 +1,11 @@
 {
-  "extends": "@tsconfig/react-native/tsconfig.json"
+	"extends": "@tsconfig/react-native/tsconfig.json",
+	"compilerOptions": {
+		"rootDir": "src",
+		"baseUrl": ".",
+		"paths": {
+			"src/*": ["src/*"],
+			"components/*": ["src/components/*"]
+		}
+	}
 }
